NICE-502 Merge spike work. Move component registration into initStoryblok phase - import statically.
