+++ web/src/utils/initStoryblok.ts
@@ -2,99 +2,74 @@ import { apiPlugin, storyblokInit } from "@storyblok/react";
 
 import { publicRuntimeConfig } from "@/config";
 import { logger } from "@/logger";
-import dynamic from "next/dynamic";
 
-// const commonComponents = {
-// 		cardGrid: dynamic(() =>
-// 			import(/* webpackChunkName: "common-card-grid" */ "@/components/Storyblok/BasicCardGrid/BasicCardGrid").then((mod) => mod.BasicCardGrid)
-// 		),
-// 		metadata: dynamic(() =>
-// 			import(/* webpackChunkName: "common-metadata" */ "@/components/Storyblok/Metadata/Metadata").then((mod) => mod.Metadata)
-// 		),
-// 		pageHeader: dynamic(() =>
-// 			import(/* webpackChunkName: "common-page-header" */ "@/components/Storyblok/StoryblokPageHeader/StoryblokPageHeader").then((mod) => mod.StoryblokPageHeader)
-// 		),
+import { BasicCardGrid } from "@/components/Storyblok/BasicCardGrid/BasicCardGrid";
+import { Blockquote } from "@/components/Storyblok/Blockquote/Blockquote";
+import { CardGridSection } from "@/components/Storyblok/CardGridSection/CardGridSection";
+import { CardListSection } from "@/components/Storyblok/CardListSection/CardListSection";
+import { CategoryLandingPage } from "@/components/Storyblok/CategoryLandingPage/CategoryLandingPage";
+import { CategoryNavigation } from "@/components/Storyblok/CategoryNavigation/CategoryNavigation";
+import { InfoPage } from "@/components/Storyblok/InfoPage/InfoPage";
+import { Metadata } from "@/components/Storyblok/Metadata/Metadata";
+import { NestedRichText } from "@/components/Storyblok/NestedRichText/NestedRichText";
+import { PromoBox } from "@/components/Storyblok/PromoBox/PromoBox";
+import { StoryblokAccordion } from "@/components/Storyblok/StoryblokAccordion/StoryblokAccordion";
+import { StoryblokAccordionGroup } from "@/components/Storyblok/StoryblokAccordionGroup/StoryblokAccordionGroup";
+import { StoryblokActionBannerDefault } from "@/components/Storyblok/StoryblokActionBanner/StoryblokActionBannerDefault";
+import { StoryblokActionBannerFullWidth } from "@/components/Storyblok/StoryblokActionBanner/StoryblokActionBannerFullWidth";
+import { StoryblokCalloutCard } from "@/components/Storyblok/StoryblokCalloutCard/StoryblokCalloutCard";
+import { StoryblokHero } from "@/components/Storyblok/StoryblokHero/StoryblokHero";
+import { StoryblokIframe } from "@/components/Storyblok/StoryblokIframe/StoryblokIframe";
+import { StoryblokPageHeader } from "@/components/Storyblok/StoryblokPageHeader/StoryblokPageHeader";
+import { StoryblokTable } from "@/components/Storyblok/StoryblokTable/StoryblokTable";
+import { StoryblokTestimonialFullWidth } from "@/components/Storyblok/StoryblokTestimonialFullWidth/StoryblokTestimonialFullWidth";
+import { StoryblokTestimonialGridItem } from "@/components/Storyblok/StoryblokTestimonialGridItem/StoryblokTestimonialGridItem";
+import { StoryblokYoutubeEmbed } from "@/components/Storyblok/StoryblokYoutubeEmbed/StoryblokYoutubeEmbed";
 
-// 		categoryLandingPage: dynamic(() =>
-// 			import(/* WebpackChunkName: "category-landing-page" */ "@/components/Storyblok/CategoryLandingPage/CategoryLandingPage").then((mod) => mod.CategoryLandingPage)
-// 		),
-// 		hero: dynamic(() =>
-// 			import(/* WebpackChunkName: "category-landing-hero" */ "@/components/Storyblok/StoryblokHero/StoryblokHero").then((mod) => mod.StoryblokHero)
-// 		),
-// 		actionBanner: dynamic(() =>
-// 			import(/* WebpackChunkName: "category-landing-action-banner" */ "@/components/Storyblok/StoryblokActionBanner/StoryblokActionBannerFullWidth").then((mod) => mod.StoryblokActionBannerFullWidth)
-// 		),
-// 		actionBannerDefault: dynamic(() =>
-// 			import(/* WebpackChunkName: "category-landing-action-banner-default" */ "@/components/Storyblok/StoryblokActionBanner/StoryblokActionBannerDefault").then((mod) => mod.StoryblokActionBannerDefault)
-// 		),
-// 		cardGridSection: dynamic(() =>
-// 			import(/* WebpackChunkName: "category-landing-card-grid-section" */ "@/components/Storyblok/CardGridSection/CardGridSection").then((mod) => mod.CardGridSection)
-// 		),
-// 		cardListSection: dynamic(() =>
-// 			import(/* WebpackChunkName: "category-landing-card-list-section" */ "@/components/Storyblok/CardListSection/CardListSection").then((mod) => mod.CardListSection)
-// 		),
-// 		promoBox: dynamic(() =>
-// 			import(/* WebpackChunkName: "category-landing-promo-box" */ "@/components/Storyblok/PromoBox/PromoBox").then((mod) => mod.PromoBox)
-// 		),
-// 		calloutCard: dynamic(() =>
-// 			import(/* WebpackChunkName: "category-landing-callout-card" */ "@/components/Storyblok/StoryblokCalloutCard/StoryblokCalloutCard").then((mod) => mod.StoryblokCalloutCard)
-// 		),
-// 		calloutCardWithImage: dynamic(() =>
-// 			import(/* WebpackChunkName: "category-landing-callout-card-with-image" */ "@/components/Storyblok/StoryblokCalloutCard/StoryblokCalloutCard").then((mod) => mod.StoryblokCalloutCard)
-// 		),
-// 		testimonialFullWidth: dynamic(() =>
-// 			import(/* WebpackChunkName: "category-landing-testimonial-full-width" */ "@/components/Storyblok/StoryblokTestimonialFullWidth/StoryblokTestimonialFullWidth").then((mod) => mod.StoryblokTestimonialFullWidth)
-// 		),
-// 		testimonialGridItem: dynamic(() =>
-// 			import(/* WebpackChunkName: "category-landing-testimonial-grid-item" */ "@/components/Storyblok/StoryblokTestimonialGridItem/StoryblokTestimonialGridItem").then((mod) => mod.StoryblokTestimonialGridItem)
-// 		),
+const commonComponents = {
+        cardGrid: BasicCardGrid,
+        metadata: Metadata,
+        pageHeader: StoryblokPageHeader,
+    };
 
-// 		accordion: dynamic(() =>
-// 			import(/* WebpackChunkName: "accordion" */ "@/components/Storyblok/StoryblokAccordion/StoryblokAccordion").then((mod) => mod.StoryblokAccordion)
-// 		),
-// 		accordionGroup: dynamic(() =>
-// 			import(/* WebpackChunkName: "accordion-group" */ "@/components/Storyblok/StoryblokAccordionGroup/StoryblokAccordionGroup").then((mod) => mod.StoryblokAccordionGroup)
-// 		),
+    //TODO: add the rest of the components as we iterate through the page build
+    const categoryLandingPageComponents = {
+        categoryLandingPage: CategoryLandingPage,
+        hero: StoryblokHero,
+        actionBanner: StoryblokActionBannerFullWidth,
+        actionBannerDefault: StoryblokActionBannerDefault,
+        cardGridSection: CardGridSection,
+        cardListSection: CardListSection,
+        promoBox: PromoBox,
+        calloutCard: StoryblokCalloutCard,
+        calloutCardWithImage: StoryblokCalloutCard,
+        testimonialFullWidth: StoryblokTestimonialFullWidth,
+        testimonialGridItem: StoryblokTestimonialGridItem,
+    };
 
-// 		iframe: dynamic(() =>
-// 			import(/* WebpackChunkName: "iframe" */ "@/components/Storyblok/StoryblokIframe/StoryblokIframe").then((mod) => mod.StoryblokIframe)
-// 		),
-// 		infoPage: dynamic(() =>
-// 			import(/* WebpackChunkName: "info-page" */ "@/components/Storyblok/InfoPage/InfoPage").then((mod) => mod.InfoPage)
-// 		),
-// 		nestedRichText: dynamic(() =>
-// 			import(/* WebpackChunkName: "nested-rich-text" */ "@/components/Storyblok/NestedRichText/NestedRichText").then((mod) => mod.NestedRichText)
-// 		),
-// 		quote: dynamic(() =>
-// 			import(/* WebpackChunkName: "blockquote" */ "@/components/Storyblok/Blockquote/Blockquote").then((mod) => mod.Blockquote)
-// 		),
-// 		youtubeEmbed: dynamic(() =>
-// 			import(/* WebpackChunkName: "youtube-embed" */ "@/components/Storyblok/StoryblokYoutubeEmbed/StoryblokYoutubeEmbed").then((mod) => mod.StoryblokYoutubeEmbed)
-// 		),
-// 		nestedTable: dynamic(() =>
-// 			import(/* WebpackChunkName: "nested-table" */ "@/components/Storyblok/StoryblokTable/StoryblokTable").then((mod) => mod.StoryblokTable)
-// 		),
-// 		categoryNavigation: dynamic(() =>
-// 						import(
-// 							/* webpackChunkName: "category-navigation" */ "@/components/Storyblok/CategoryNavigation/CategoryNavigation"
-// 						).then((mod) => mod.CategoryNavigation)
-// 					),
-// 	};
+    const infoPageComponents = {
+        accordion: StoryblokAccordion,
+        accordionGroup: StoryblokAccordionGroup,
+        hero: StoryblokHero,
+        iframe: StoryblokIframe,
+        infoPage: InfoPage,
+        nestedRichText: NestedRichText,
+        quote: Blockquote,
+        youtubeEmbed: StoryblokYoutubeEmbed,
+        actionBannerDefault: StoryblokActionBannerDefault,
+        nestedTable: StoryblokTable,
+    };
 
-// 	// const components = {
-// 	// 	...commonComponents,
-// 	// 	...(component === "infoPage"
-// 	// 		? infoPageComponents
-// 	// 		: component === "categoryLandingPage"
-// 	// 		? categoryLandingPageComponents
-// 	// 		: {
-// 	// 				categoryNavigation: dynamic(() =>
-// 	// 					import(
-// 	// 						/* webpackChunkName: "category-navigation" */ "@/components/Storyblok/CategoryNavigation/CategoryNavigation"
-// 	// 					).then((mod) => mod.CategoryNavigation)
-// 	// 				),
-// 	// 		  }),
-// 	// };
+	const categoryNavigationPageComponents = {
+		categoryNavigation: CategoryNavigation
+	}
+
+	const allComponents = {
+		...commonComponents,
+		...infoPageComponents,
+		...categoryLandingPageComponents,
+		...categoryNavigationPageComponents
+	}
 
 
 // Init connection to Storyblok
@@ -113,6 +88,7 @@ export const initStoryblok = (): void => {
 		storyblokInit({
 			accessToken,
 			use: [apiPlugin],
+			components: allComponents,
 			apiOptions: {
 				cache: {
 					clear: usingOcelotCache ? "manual" : "auto",
